Refactored code to eliminate repetition, significant reduction of lines!
With the guidance of Recurser Robert Xu, I was able to create two new functions, moveUp and moveDown, which each take parameters compFist and playerFist. along with the container variable and new timerUp and timerDown variables. This made it so that I could call the two functions within my shake function and replace the variables with the specific times.

Robert recommended that my next step be to create a loop to implement the times, calling moveUp and moveDown less times, and reducing the number of lines further!